requirements.txt
torch
torchvision
numpy
pillow
h5py
nltk
matplotlib
tqdm
pandas
